// products/pc-admin/packages/shared-core/src/btc/plugins/manager/types.d.ts
export interface PluginMetadata {
    /**
     * 鎻掍欢鏄剧ず鍚嶇О
     */
    label?: string;
    /**
     * 鎻掍欢鎻忚堪
     */
    description?: string;
    /**
     * 浣滆€?   */
    author?: string;
    /**
     * 鐗堟湰鍙?   */
    version?: string;
    /**
     * 鏇存柊鏃堕棿
     */
    updateTime?: string;
    /**
     * 鎻掍欢鏂囨。閾炬帴
     */
    doc?: string;
    /**
     * 绀轰緥璺緞鍒楄〃
     */
    demo?: string[];
    /**
     * 鎻掍欢鍥炬爣 URL
     */
    icon?: string;
    /**
     * 鎻掍欢鍒嗙被
     */
    category?: string;
    /**
     * 鎻掍欢鏍囩
     */
    tags?: string[];
    /**
     * 鏄惁鎺ㄨ崘浣跨敤
     */
    recommended?: boolean;
}